// styled_components/src/styledComponents/types.ts
export interface Theme {
    colors: typeof colors;
    dimensions: typeof dimensions;
    fonts: typeof fonts;
}

export type TColorThemeHook = {
    colors?: typeof colors;
};

export type ThemeElement = {
    theme?: Theme;
};

export type TColorAnimate = { 
    bgColor?: ColorType;
    bgMiddleColor?: ColorType;
    count?: number | string;
    isSearch?: boolean;
};
export type ColorType1 = typeof colors;
export type ColorType = keyof ColorType1;
export type IconColorType = 'inherit' | 'primary' | 'secondary' | 'action' | 'disabled' | 'error';
export interface ColorsProps {
    backgroundColor?: ColorType;
    color?: ColorType;
    brColor?: ColorType;
    phColor?: ColorType;
    hwColor?: ColorType;
    hwBgColor?: ColorType;
    brColorTop?: ColorType;
    brColorLeft?: ColorType;
    brColorBottom?: ColorType;
    opacity?: number;
}
// ...
export interface SizesProps {
    width?: string | number;
    width1366?: string | number;
    padding?: string | number;
    margin?: string | number;
    top?: string | number;
    left?: string | number;
    height?: string | number;
    maxHeight?: string | number;
    bxSizing?: string;
    minWidth?: string | number;
    maxWidth?: string | number;
    marginBottom?: string | number;
    minHeight?: string | number;
}
// ...
export interface StCellProps extends SizesProps {
    withoutTSError?: any;
}